rename use-put-task-completion
